added export html functionality to irg archive document view
